server/requirements.txt
openenv-core[core]>=0.2.2
pydantic>=2.0.0
fastapi>=0.115.0
uvicorn>=0.24.0
